code/requirements.txt
Django==2.2.1
django-rest-auth==0.9.5
django-rest-knox==4.0.1
djangorestframework==3.9.3
mysqlclient==1.4.2
oauthlib==1.0.3
Pillow==7.1.2
prettytable==0.7.2
python-decouple==3.3
requests==2.9.1
six==1.10.0
sqlparse==0.3.1
urllib3==1.13.1
